requirements.txt
streamlit
pandas
numpy
plotly
altair
matplotlib
python-dotenv
transformers
torch
requests
